Move subscription types to dedicated file
Refactored Zuora subscription schemas and types from zuoraSchemas.ts to a new file, subscription.ts, under types/objects. This improves code organization and separation of concerns.
