+++ modules/zuora/src/types/objects/subscription.ts
@@ -1,0 +1,58 @@
+import { z } from 'zod';
+import { BillingPeriodValues } from '@modules/billingPeriod';
+import { zuoraResponseSchema } from '../httpResponse';
+
+export const zuoraSubscriptionSchema = z.object({
+	id: z.string(),
+	accountNumber: z.string(),
+	subscriptionNumber: z.string(),
+	status: z.string(),
+	contractEffectiveDate: z.coerce.date(),
+	serviceActivationDate: z.coerce.date(),
+	customerAcceptanceDate: z.coerce.date(),
+	subscriptionStartDate: z.coerce.date(),
+	subscriptionEndDate: z.coerce.date(),
+	lastBookingDate: z.coerce.date(),
+	termStartDate: z.coerce.date(),
+	termEndDate: z.coerce.date(),
+	ratePlans: z.array(
+		z.object({
+			id: z.string(),
+			lastChangeType: z.optional(z.string()),
+			productId: z.string(),
+			productName: z.string(),
+			productRatePlanId: z.string(),
+			ratePlanName: z.string(),
+			ratePlanCharges: z.array(
+				z.object({
+					id: z.string(),
+					productRatePlanChargeId: z.string(),
+					number: z.string(),
+					name: z.string(),
+					type: z.string(),
+					model: z.string(),
+					currency: z.string(),
+					effectiveStartDate: z.coerce.date(),
+					effectiveEndDate: z.coerce.date(),
+					billingPeriod: z.nullable(z.enum(BillingPeriodValues)),
+					processedThroughDate: z.coerce.date(),
+					chargedThroughDate: z.coerce.date().nullable(),
+					upToPeriodsType: z.nullable(z.string()),
+					upToPeriods: z.nullable(z.number()),
+					price: z.nullable(z.number()),
+					discountPercentage: z.nullable(z.number()),
+				}),
+			),
+		}),
+	),
+});
+
+export const zuoraSubscriptionResponseSchema = zuoraResponseSchema.merge(
+	zuoraSubscriptionSchema,
+);
+
+export type ZuoraSubscription = z.infer<typeof zuoraSubscriptionSchema>;
+
+export type RatePlan = ZuoraSubscription['ratePlans'][number];
+
+export type RatePlanCharge = RatePlan['ratePlanCharges'][number];
