+++ modules/zuora/src/zuoraSchemas.ts
@@ -18,63 +18,6 @@ export const zuoraBearerTokenSchema = z.object({
 	expires_in: z.number(),
 });
 
-// --------------- Subscription ---------------
-export const zuoraSubscriptionSchema = z.object({
-	id: z.string(),
-	accountNumber: z.string(),
-	subscriptionNumber: z.string(),
-	status: z.string(),
-	contractEffectiveDate: z.coerce.date(),
-	serviceActivationDate: z.coerce.date(),
-	customerAcceptanceDate: z.coerce.date(),
-	subscriptionStartDate: z.coerce.date(),
-	subscriptionEndDate: z.coerce.date(),
-	lastBookingDate: z.coerce.date(),
-	termStartDate: z.coerce.date(),
-	termEndDate: z.coerce.date(),
-	ratePlans: z.array(
-		z.object({
-			id: z.string(),
-			lastChangeType: z.optional(z.string()),
-			productId: z.string(),
-			productName: z.string(),
-			productRatePlanId: z.string(),
-			ratePlanName: z.string(),
-			ratePlanCharges: z.array(
-				z.object({
-					id: z.string(),
-					productRatePlanChargeId: z.string(),
-					number: z.string(),
-					name: z.string(),
-					type: z.string(),
-					model: z.string(),
-					currency: z.string(),
-					effectiveStartDate: z.coerce.date(),
-					effectiveEndDate: z.coerce.date(),
-					billingPeriod: z.nullable(z.enum(BillingPeriodValues)),
-					processedThroughDate: z.coerce.date(),
-					chargedThroughDate: z.coerce.date().nullable(),
-					upToPeriodsType: z.nullable(z.string()),
-					upToPeriods: z.nullable(z.number()),
-					price: z.nullable(z.number()),
-					discountPercentage: z.nullable(z.number()),
-				}),
-			),
-		}),
-	),
-});
-export const zuoraSubscriptionResponseSchema = z
-	.object({
-		success: z.boolean(),
-	})
-	.merge(zuoraSubscriptionSchema);
-
-export type ZuoraSubscription = z.infer<typeof zuoraSubscriptionSchema>;
-
-export type RatePlan = ZuoraSubscription['ratePlans'][number];
-
-export type RatePlanCharge = RatePlan['ratePlanCharges'][number];
-
 // --------------- Account ---------------
 export const zuoraAccountBasicInfoSchema = z
 	.object({
